requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
torch==2.1.1
torchvision==0.16.1
Pillow==10.1.0
matplotlib==3.8.2
numpy==1.24.3
PyYAML==6.0.1 